+++ db/migrate/20150216213546_add_bdate_to_users.rb
@@ -1,0 +1,8 @@
+class AddBdateToUsers < ActiveRecord::Migration
+  def self.up
+    add_column :users, :bdate, :date
+  end
+  def self.down
+    remove_column :users, :bdate
+  end
+end